fix: widen TurnCost token fields to u64 and remove unused variable
- Change input_tokens/output_tokens from u32 to u64 in StatusUpdate::TurnCost,
  SseEvent::TurnCost, and the thread_ops emit site to avoid truncation on
  large conversations
- Remove unused _routine_engine_for_loop binding in agent_loop.rs

+++ src/agent/thread_ops.rs
@@ -516,8 +516,8 @@ impl Agent {
                         .send_status(
                             &message.channel,
                             StatusUpdate::TurnCost {
-                                input_tokens: total_in as u32,
-                                output_tokens: total_out as u32,
+                                input_tokens: total_in,
+                                output_tokens: total_out,
                                 cost_usd: format!("${:.4}", total_cost),
                             },
                             &message.metadata,
